+++ src/yamk/main.py
@@ -1,10 +1,15 @@
 import argparse
 import sys
+import warnings
 
 from yamk import __version__
 from yamk.make import MakeCommand
 
 sys.tracebacklimit = 0
+
+
+class RemovedInYam3(Warning):
+    pass
 
 
 def parse_args():
@@ -29,12 +34,18 @@ def parse_args():
         action="store_true",
         help="rebuild all dependencies and the target",
     )
-    parser.add_argument(
+    group = parser.add_mutually_exclusive_group()
+    group.add_argument(
         "-m",
         "--makefile",
         metavar="Makefile",
-        default="make.toml",
-        help="the path to makefile",
+        help="the path to the makefile",
+    )
+    group.add_argument(
+        "-c",
+        "--cookbook",
+        metavar="cookbook",
+        help="the path to the cookbook",
     )
     parser.add_argument(
         "-s",
@@ -72,4 +83,14 @@ def parse_args():
 
 def main():
     args = parse_args()
+    if args.makefile is not None:
+        warnings.warn(
+            "Using -m/--makefile is deprecated, and will be removed in yamk 2.0. "
+            "Use -c/--cookbook instead.",
+            RemovedInYam3,
+        )
+        args.cookbook = args.makefile
+        args.makefile = None
+    elif args.cookbook is None:
+        args.cookbook = "make.toml"
     MakeCommand(args).make()